requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.0
bidict==0.23.1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
docutils==0.21.2
fastapi==0.115.12
frozenlist==1.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
h11==0.16.0
htmx==0.0.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown2==2.5.3
MarkupSafe==3.0.2
multidict==6.4.4
nicegui==2.20.0
oauthlib==3.2.2
orjson==3.10.18
propcache==0.3.2
pscript==0.7.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.22
Pygments==2.19.1
python-dotenv==1.1.0
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
vbuild==0.8.2
watchfiles==1.1.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
